changed state initialization to layout file
